benchmarking/requirements.txt
# Additional requirements for benchmarking
# Core dependencies (should already be installed)
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
pandas>=1.5.0
Pillow>=9.0.0

# VLM dependencies
transformers>=4.30.0  # For LLaVA
accelerate>=0.20.0  # For model loading

# Chess analysis dependencies
chess>=1.999  # Python chess library (pychess)
requests>=2.28.0  # For Lichess API

# Optional but recommended
numpy>=1.24.0
tqdm>=4.65.0

